Adicionando Controlador de Build

+++ src/main/java/com/adi/tftapi/Repository/BuildRepository.java
@@ -3,5 +3,5 @@ package com.adi.tftapi.Repository;
 import com.adi.tftapi.Entity.Build;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BuildRepository extends JpaRepository<Build, Long> {
+public interface BuildRepository extends JpaRepository<Build, Integer> {
 }
